- Добавлены классы исключений - Произведён рефакторинг репозиториев(integer id -> long id) - Теперь в виде ответа при некорретных запросах с id приходит ошибка и её детали - Postgres миграция обновлена (IDENTITY -> BIGSERIAL) - Изменены Docker-compose (Теперь поднимается только image Postgre)

+++ study/src/main/java/ru/darnei/study/repository/DataStorageJdbcTemplateImplementation.java
@@ -3,11 +3,13 @@ package ru.darnei.study.repository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import ru.darnei.study.exception.ResourceNotFoundException;
 import ru.darnei.study.model.User;
 
 import java.sql.PreparedStatement;
@@ -31,10 +33,10 @@ class DataStorageJdbcTemplateImplementation implements DataStorage{
     }
 
     @Override
-    public User updateUser(Integer id, User user) {
+    public ResponseEntity<User> updateUser(Long id, User user) throws ResourceNotFoundException {
         String sql = "SELECT COUNT(*) FROM usr WHERE id = ?";
         boolean exists = false;
-        int count = jdbcTemplate.queryForObject(sql,new Object[]{id}, Integer.class);
+        long count = jdbcTemplate.queryForObject(sql,new Object[]{id}, Long.class);
         exists = count > 0;
         if(exists) {
             jdbcTemplate.update("UPDATE USR SET login=?, email=?, password=?, salary=? WHERE id=?",
@@ -43,17 +45,18 @@ class DataStorageJdbcTemplateImplementation implements DataStorage{
                     user.getPassword(),
                     user.getSalary(),
                     id);
-            return new User(id.longValue(), user.getLogin(), user.getEmail(), user.getPassword(), user.getSalary());
+            return ResponseEntity.ok(new User(id.longValue(), user.getLogin(), user.getEmail(), user.getPassword(), user.getSalary()));
         }
         else
-            return null;
+            throw new ResourceNotFoundException("User not found for this id : " + id);
     }
 
 
 
     @Override
     public List<User> findAllUsers() {
-        List<User> users = jdbcTemplate.query("SELECT * FROM usr", (result, rowNum) -> new User(
+        List<User> users = jdbcTemplate.query("SELECT * FROM usr", (result, rowNum)
+                -> new User(
                 result.getLong("id"),
                 result.getString("login"),
                 result.getString("email"),
@@ -63,13 +66,13 @@ class DataStorageJdbcTemplateImplementation implements DataStorage{
     }
 
     @Override
-    public User findByIdUser(Integer id) {
-        User user = jdbcTemplate.queryForObject("SELECT * FROM usr WHERE ID=?", new Object[]{id}, new BeanPropertyRowMapper<>(User.class));
-        return user;
+    public ResponseEntity<User> findByIdUser(Long id) throws ResourceNotFoundException {
+            User user = jdbcTemplate.queryForObject("SELECT * FROM usr WHERE ID=?", new Object[]{id}, new BeanPropertyRowMapper<>(User.class));
+            return ResponseEntity.ok(user);
     }
 
     @Override
-    public User newUser(User user) {
+    public ResponseEntity<User> newUser(User user) {
         String sql = "INSERT INTO usr (login,email,password,salary) VALUES (?,?,?,? )";
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update(connection -> {
@@ -80,11 +83,20 @@ class DataStorageJdbcTemplateImplementation implements DataStorage{
             ps.setInt(4, user.getSalary());
             return ps;
         }, keyHolder);
-        return new User(keyHolder.getKey().longValue(), user.getLogin(), user.getEmail(), user.getPassword(), user.getSalary());
+       return ResponseEntity.ok(new User(keyHolder.getKey().longValue(), user.getLogin(), user.getEmail(), user.getPassword(), user.getSalary()));
     }
 
     @Override
-    public void deleteUser(Integer id) {
-        jdbcTemplate.update("DELETE FROM usr WHERE id=?", id);
-    }
+    public void deleteUser(Long id) throws ResourceNotFoundException {
+        String sql = "SELECT COUNT(*) FROM usr WHERE id = ?";
+        boolean exists = false;
+        long count = jdbcTemplate.queryForObject(sql,new Object[]{id}, Long.class);
+        exists = count > 0;
+        if(exists) {
+            jdbcTemplate.update("DELETE FROM usr WHERE id=?", id);
+        }
+        else
+            throw new ResourceNotFoundException("User not found for this id : " + id);
+        }
+
 }
